+++ HISTORY.rst
@@ -26,3 +26,10 @@ History
 ------------------
 
 * Fixed error where exception thrown as tests cleaned up.
+
+0.4.0 (2020-10-06)
+------------------
+
+* Split out handling in API of Types and Commands.
+* This changes ``get_command_class()`` to ``get_type_class()`` and adds
+  ``get_type_object()``